abstract out most of tables info from grib2 class - this would make it easier to maintain and add local tabels thanks to javascript language
