+++ nf0.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = function nf0(number) {
+    return Math.round(number);
+};
